% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_helpers.R
\name{alt_extractor}
\alias{alt_extractor}
\title{Alternate extractor that expects a slightly different structure}
\usage{
alt_extractor(x)
}
\arguments{
\item{x}{A parsed value (likely from a parser combinator)}
}
\description{
Alternate extractor that expects a slightly different structure
}
\keyword{internal}
