% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{anat_spec}
\alias{anat_spec}
\title{Create a spec table for "anat" files}
\usage{
anat_spec()
}
\value{
A list containing \code{keystruc}, \code{kinds}, and \code{type} describing
anatomical files.
}
\description{
Create a spec table for "anat" files
}
\examples{
anat_spec()
}
\keyword{internal}
