% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{anatprepspec}
\alias{anatprepspec}
\title{Create a spec table for fMRIPrep "anat" files}
\usage{
anatprepspec()
}
\description{
Create a spec table for fMRIPrep "anat" files
}
\value{
A list containing \code{keystruc}, \code{kinds}, and \code{type} for fMRIPrep anatomical files.
}
\examples{
anatprepspec()
}
\keyword{internal}
