% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{bids_project}
\alias{bids_project}
\title{Load a BIDS project at a file path}
\usage{
bids_project(path = ".", fmriprep = FALSE, prep_dir = "derivatives/fmriprep")
}
\arguments{
\item{path}{the file path of the project}
}
\description{
bids project
}
\examples{

# p <- system.file("inst/extdata/7t_trt", package="bidser")
# path <- "~/code/bidser/inst/extdata/7t_trt"
# pp <- bids_project(p)

# pp2 <- bids_project(system.file("inst/extdata/megalocalizer", package="bidser"), fmriprep=TRUE)
}
