% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_uri.R
\name{bids_uri}
\alias{bids_uri}
\title{Construct a BIDS URI object}
\usage{
bids_uri(uri)
}
\arguments{
\item{uri}{A character scalar BIDS URI, e.g.
\code{"bids::sub-01/fmap/sub-01_epi.nii.gz"} or
\code{"bids:deriv1:sub-01/anat/T1w.nii.gz"}.}
}
\value{
An object of class \code{bids_uri} with fields \code{dataset_name},
\code{relative_path}, and \code{uri}.
}
\description{
Parses a BIDS URI string of the form \verb{bids:<dataset_name>:<relative_path>}
and returns an S3 object of class \code{bids_uri}.
}
\examples{
u <- bids_uri("bids::sub-01/func/sub-01_task-rest_bold.nii.gz")
u$dataset_name   # ""
u$relative_path  # "sub-01/func/sub-01_task-rest_bold.nii.gz"

# URI with a named dataset link
u2 <- bids_uri("bids:deriv1:sub-01/anat/sub-01_T1w.nii.gz")
u2$dataset_name  # "deriv1"
}
