% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids.R,
%   R/dataset_description.R
\name{bids_version}
\alias{bids_version}
\alias{bids_version.bids_project}
\alias{bids_version.mock_bids_project}
\alias{bids_version.bids_dataset_description}
\title{Get the BIDS version of a dataset}
\usage{
bids_version(x, ...)

\method{bids_version}{bids_project}(x, ...)

\method{bids_version}{mock_bids_project}(x, ...)

\method{bids_version}{bids_dataset_description}(x, ...)
}
\arguments{
\item{x}{A \code{bids_dataset_description} or \code{bids_project} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character scalar BIDS version string, or \code{NA_character_}.
}
\description{
Get the BIDS version of a dataset
}
\examples{
\donttest{
tryCatch({
  ds001_path <- get_example_bids_dataset("ds001")
  desc <- read_dataset_description(ds001_path)
  bids_version(desc)
  unlink(ds001_path, recursive = TRUE)
}, error = function(e) message("Example requires internet: ", e$message))
}
}
