% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_func_scans}
\alias{check_func_scans}
\title{x <- bids_project(system.file("inst/extdata/phoneme_stripped", package="bidser"), fmriprep=TRUE)}
\usage{
check_func_scans(x)
}
\description{
x <- bids_project(system.file("inst/extdata/phoneme_stripped", package="bidser"), fmriprep=TRUE)
}
