% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_bids.R
\name{create_preproc_mask.mock_bids_project}
\alias{create_preproc_mask.mock_bids_project}
\title{Create Preprocessing Mask (Mock Implementation)}
\usage{
\method{create_preproc_mask}{mock_bids_project}(x, ...)
}
\arguments{
\item{x}{A \code{mock_bids_project} object.}

\item{...}{Arguments (ignored).}
}
\value{
Throws an error indicating the function is not applicable to mock objects.
}
\description{
This function is not implemented for \code{mock_bids_project} objects as they do not
contain actual image data required to create a mask.
}
\examples{
mock <- create_mock_bids("Test", c("01"), tibble::tibble(
  subid = "01", datatype = "func",
  suffix = "bold.nii.gz", fmriprep = FALSE
))
try(create_preproc_mask(mock))
}
