% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{create_preproc_mask}
\alias{create_preproc_mask}
\title{create pre-processing mask}
\usage{
create_preproc_mask(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
create a binary mask from pre-processed functional scans
}
