% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_parsers.R
\name{create_regex_parser}
\alias{create_regex_parser}
\title{Create a regex-based parser from a specification}
\usage{
create_regex_parser(spec)
}
\arguments{
\item{spec}{A specification object}
}
\value{
A parser function
}
\description{
Create a regex-based parser from a specification
}
\keyword{internal}
