% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{decode}
\alias{decode}
\title{Decode a key-value list into a string}
\usage{
decode(x, ...)
}
\arguments{
\item{x}{the list to decode}

\item{...}{extra args}
}
\description{
Decode a key-value list into a string
}
