% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{event_files}
\alias{event_files}
\title{Get event files}
\usage{
event_files(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
get vector of event files in project
}
