% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_parsers.R
\name{extract_bids_entities}
\alias{extract_bids_entities}
\title{Extract BIDS entities from the entity part of filename}
\usage{
extract_bids_entities(entity_part, keystruc_spec)
}
\arguments{
\item{entity_part}{The part of filename before the kind}

\item{keystruc_spec}{The key structure specification}
}
\value{
Named list of entities or NULL
}
\description{
Extract BIDS entities from the entity part of filename
}
\keyword{internal}
