% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_parsers.R
\name{extract_kind_and_suffix}
\alias{extract_kind_and_suffix}
\title{Extract kind and suffix from filename}
\usage{
extract_kind_and_suffix(filename, kinds_spec)
}
\arguments{
\item{filename}{The filename}

\item{kinds_spec}{The kinds specification}
}
\value{
List with kind and suffix or NULL
}
\description{
Extract kind and suffix from filename
}
\keyword{internal}
