% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmap_parser}
\alias{fmap_parser}
\title{construct a parser for field map types}
\usage{
fmap_parser()
}
\description{
construct a parser for field map types
}
\examples{
p <- fmap_parser()
parse(p, "sub-01_ses-1_run-1_magnitude1.nii.gz")
parse(p, "sub-01_ses-1_run-1_phasediff.nii.gz")
parse(p, "sub-01_ses-1_run-2_magnitude2.nii.gz")
}
