% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmriprep_anat_parser}
\alias{fmriprep_anat_parser}
\title{fMRIPrep anatomical parser constructor}
\usage{
fmriprep_anat_parser()
}
\value{
An fMRIPrep anatomical parser object for parsing preprocessed anatomical files
}
\description{
fMRIPrep anatomical parser constructor
}
\examples{
# Create an fMRIPrep anatomical parser
parser <- fmriprep_anat_parser()

# Parse a preprocessed anatomical file
result <- parse(parser, "sub-01_space-MNI152NLin2009cAsym_desc-preproc_T1w.nii.gz")
}
