% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmriprep_func_parser}
\alias{fmriprep_func_parser}
\title{fMRIPrep functional parser constructor}
\usage{
fmriprep_func_parser()
}
\value{
An fMRIPrep functional parser object for parsing preprocessed functional files
}
\description{
fMRIPrep functional parser constructor
}
\examples{
# Create an fMRIPrep functional parser
parser <- fmriprep_func_parser()

# Parse a preprocessed functional file
result <- parse(parser, "sub-01_task-rest_space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz")
}
