% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{func_parser}
\alias{func_parser}
\title{Functional parser constructor}
\usage{
func_parser()
}
\value{
A functional BIDS parser object for parsing functional MRI files
}
\description{
Functional parser constructor
}
\examples{
# Create a functional parser
parser <- func_parser()

# Parse a functional file
result <- parse(parser, "sub-01_task-rest_run-01_bold.nii.gz")
}
