% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_uri.R
\name{is_bids_uri}
\alias{is_bids_uri}
\title{Test whether an object is a BIDS URI}
\usage{
is_bids_uri(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"bids_uri"}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a BIDS URI
}
\examples{
u <- bids_uri("bids::sub-01/anat/sub-01_T1w.nii.gz")
is_bids_uri(u)      # TRUE
is_bids_uri("bids::sub-01/anat/sub-01_T1w.nii.gz")  # FALSE
}
