% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_helpers.R
\name{mandatory_key}
\alias{mandatory_key}
\title{create parser for a mandatory BIDS key}
\usage{
mandatory_key(label, regex = "[A-Za-z0-9]+")
}
\description{
create parser for a mandatory BIDS key
}
\keyword{internal}
