% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_helpers.R
\name{one_of}
\alias{one_of}
\title{Match one of several possible labels}
\usage{
one_of(labels)
}
\arguments{
\item{labels}{A character vector of possible labels.}
}
\value{
A parser that matches \verb{_<one_of_labels>}
}
\description{
Given a vector of labels, matches \verb{_<label>} where \if{html}{\out{<label>}} is one of them.
}
\keyword{internal}
