% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bids.R
\name{plot_bids_completeness}
\alias{plot_bids_completeness}
\title{Plot BIDS data completeness}
\usage{
plot_bids_completeness(data, color_scheme = "viridis")
}
\arguments{
\item{data}{Preprocessed BIDS data from prepare_bids_data_for_plot}

\item{color_scheme}{Color scheme to use}
}
\value{
A ggplot object
}
\description{
Creates a heatmap showing data completeness across subjects and tasks.
}
\keyword{internal}
