% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bids.R
\name{plot_bids_heatmap}
\alias{plot_bids_heatmap}
\title{Plot BIDS data as a heatmap}
\usage{
plot_bids_heatmap(data, color_scheme = "viridis", highlight_missing = TRUE)
}
\arguments{
\item{data}{Preprocessed BIDS data from prepare_bids_data_for_plot}

\item{color_scheme}{Color scheme to use}

\item{highlight_missing}{Whether to highlight missing data}
}
\value{
A ggplot object
}
\description{
Plot BIDS data as a heatmap
}
\keyword{internal}
