% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bids.R
\name{plot_bids_tasks}
\alias{plot_bids_tasks}
\title{Plot BIDS task distribution}
\usage{
plot_bids_tasks(data, color_scheme = "viridis")
}
\arguments{
\item{data}{Preprocessed BIDS data from prepare_bids_data_for_plot}

\item{color_scheme}{Color scheme to use}
}
\value{
A ggplot object
}
\description{
Creates a plot showing the distribution of tasks across subjects.
}
\keyword{internal}
