% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bids.R
\name{prepare_bids_data_for_plot}
\alias{prepare_bids_data_for_plot}
\title{Prepare BIDS data for visualization}
\usage{
prepare_bids_data_for_plot(x, include_derivatives = TRUE)
}
\arguments{
\item{x}{A bids_project object}

\item{include_derivatives}{Logical. Whether to include derivatives data}
}
\value{
A list containing project info and formatted data
}
\description{
This internal function processes a bids_project object and extracts the necessary
data for visualization, including project info and formatted data.
}
\keyword{internal}
