% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{preproc_scans}
\alias{preproc_scans}
\title{Get pre-processed fmri scans}
\usage{
preproc_scans(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
extract fmriprep-created \code{preproc} scans from bids project
}
