% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{read_confounds}
\alias{read_confounds}
\title{read in confound files}
\usage{
read_confounds(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
read in confound files
}
