% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/tabulars.R
\name{read_participants}
\alias{read_participants}
\alias{read_participants.bids_project}
\alias{read_participants.character}
\title{Read participants.tsv as a typed tabular object}
\usage{
read_participants(x, ...)

\method{read_participants}{bids_project}(x, ...)

\method{read_participants}{character}(x, ...)
}
\arguments{
\item{x}{A \code{bids_project} object or a character path.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{bids_participants} object inheriting from \code{tbl_df}, or NULL.
}
\description{
Read participants.tsv as a typed tabular object
}
