% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/tabulars.R
\name{read_scans_tsv}
\alias{read_scans_tsv}
\alias{read_scans_tsv.bids_project}
\title{Read a scans.tsv file for a subject}
\usage{
read_scans_tsv(x, subid, session = NULL, ...)

\method{read_scans_tsv}{bids_project}(x, subid, session = NULL, ...)
}
\arguments{
\item{x}{A \code{bids_project} object.}

\item{subid}{Subject ID (without \verb{sub-} prefix).}

\item{session}{Optional session ID (without \verb{ses-} prefix).}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{bids_scans_tsv} object inheriting from \code{tbl_df}, or NULL.
}
\description{
Read a scans.tsv file for a subject
}
