% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/tabulars.R
\name{read_sessions_tsv}
\alias{read_sessions_tsv}
\alias{read_sessions_tsv.bids_project}
\title{Read a sessions.tsv file for a subject}
\usage{
read_sessions_tsv(x, subid, ...)

\method{read_sessions_tsv}{bids_project}(x, subid, ...)
}
\arguments{
\item{x}{A \code{bids_project} object.}

\item{subid}{Subject ID (without \verb{sub-} prefix).}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{bids_sessions_tsv} object inheriting from \code{tbl_df}, or NULL.
}
\description{
Read a sessions.tsv file for a subject
}
