% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids_uri.R
\name{resolve_bids_uri}
\alias{resolve_bids_uri}
\alias{resolve_bids_uri.bids_uri}
\alias{resolve_bids_uri.character}
\title{Resolve a BIDS URI to a local or remote path}
\usage{
resolve_bids_uri(uri, description, ..., must_exist = FALSE)

\method{resolve_bids_uri}{bids_uri}(uri, description, ..., must_exist = FALSE)

\method{resolve_bids_uri}{character}(uri, description, ..., must_exist = FALSE)
}
\arguments{
\item{uri}{A \code{bids_uri} object or a character scalar BIDS URI string.}

\item{description}{A \code{bids_dataset_description} or \code{bids_project} object.}

\item{...}{Additional arguments passed to methods.}

\item{must_exist}{Logical. If \code{TRUE}, the resolved path must exist on disk.}
}
\value{
A character scalar path (or URL for remote links).
}
\description{
Resolve a BIDS URI to a local or remote path
}
