% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/tabulars.R
\name{sidecar}
\alias{sidecar}
\alias{sidecar.bids_tabular}
\title{Get the sidecar metadata attached to a tabular BIDS object}
\usage{
sidecar(x, ...)

\method{sidecar}{bids_tabular}(x, ...)
}
\arguments{
\item{x}{A \code{bids_tabular} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of sidecar metadata, or an empty list if none.
}
\description{
Get the sidecar metadata attached to a tabular BIDS object
}
