% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_helpers.R
\name{start_key}
\alias{start_key}
\title{Create a parser for a starting BIDS key (no leading underscore)}
\usage{
start_key(label, regex = "[A-Za-z0-9]+")
}
\arguments{
\item{label}{The label of the key to parse.}

\item{regex}{The regex for the key's value. Must be a single character string.}
}
\value{
A parser that matches \verb{<label>-<id>} at the start
}
\description{
Matches \verb{<label>-<id>} at the start of a filename.
}
