% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{tasks}
\alias{tasks}
\title{Get Tasks}
\usage{
tasks(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\value{
a vector of task names
}
\description{
get the tasks in the project
}
