% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{anat_parser}
\alias{anat_parser}
\title{construct a parser for anatomical data}
\usage{
anat_parser()
}
\description{
construct a parser for anatomical data
}
\examples{
ap <- anat_parser()
parse(ap, "sub-01_ses-1_T1map.nii.gz")
parse(ap, "sub-01_ses-1_T1w.nii.gz")
parse(ap, "sub-01_ses-retest_T1w.nii.gz")

}
