% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids_uri.R
\name{as_bids_uri}
\alias{as_bids_uri}
\alias{as_bids_uri.character}
\alias{as_bids_uri.bids_uri}
\title{Coerce to a BIDS URI object}
\usage{
as_bids_uri(x, ...)

\method{as_bids_uri}{character}(x, ...)

\method{as_bids_uri}{bids_uri}(x, ...)
}
\arguments{
\item{x}{A character scalar BIDS URI string, or a \code{bids_uri} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{bids_uri} object.
}
\description{
Coerce to a BIDS URI object
}
\examples{
u <- as_bids_uri("bids::sub-01/anat/sub-01_T1w.nii.gz")
u$dataset_name   # ""
u$relative_path  # "sub-01/anat/sub-01_T1w.nii.gz"
}
