% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{bids_parser}
\alias{bids_parser}
\title{Create a parser for a generic BIDS file}
\usage{
bids_parser()
}
\description{
This parser tries to match against various known parsers (anat, func, fmriprep anat/func).
}
