% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{bids_project}
\alias{bids_project}
\title{load a BIDS project}
\usage{
bids_project(path = ".")
}
\arguments{
\item{path}{the file path of the project}
}
\description{
load a BIDS project
}
\examples{

p <- system.file("inst/extdata/ds001", package="bidser")
#path <- "~/code/bidser/inst/extdata/ds005"
}
