% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{bids_project}
\alias{bids_project}
\title{load a BIDS project}
\usage{
bids_project(path = ".", fmriprep = FALSE, prep_dir = "derivatives/fmriprep")
}
\arguments{
\item{path}{the file path of the project}
}
\description{
load a BIDS project
}
\examples{

p <- system.file("inst/extdata/7t_trt", package="bidser")
#path <- "~/code/bidser/inst/extdata/7t_trt"
pp <- bids_project(p)

pp2 <- bids_project(system.file("inst/extdata/megalocalizer", package="bidser"), fmriprep=TRUE)
}
