% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_extensions.R
\name{bids_report}
\alias{bids_report}
\title{Create a lightweight text report for a BIDS project}
\usage{
bids_report(x, ...)
}
\arguments{
\item{x}{A \code{bids_project} object.}

\item{...}{Additional arguments passed to \code{\link[=bids_report_data]{bids_report_data()}}.}
}
\value{
An object of class \code{bids_report}.
}
\description{
Create a lightweight text report for a BIDS project
}
