% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidsio.R
\name{confound_files.bids_project}
\alias{confound_files.bids_project}
\title{Locate confound files}
\usage{
\method{confound_files}{bids_project}(x, subid = ".*", task = ".*", session = ".*", ...)
}
\arguments{
\item{x}{A \code{bids_project} object}

\item{subid}{Subject ID regex}

\item{task}{Task regex}

\item{session}{Session regex}

\item{...}{Additional arguments (not currently used)}
}
\value{
A character vector of file paths
}
\description{
Locate confound files
}
