% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidsio.R
\name{create_preproc_mask.bids_project}
\alias{create_preproc_mask.bids_project}
\title{Create a preprocessing mask}
\usage{
\method{create_preproc_mask}{bids_project}(x, subid, thresh = 0.99, ...)
}
\arguments{
\item{x}{A \code{bids_project} object}

\item{subid}{Subject ID regex}

\item{thresh}{Threshold value (default 0.99)}

\item{...}{Additional arguments passed to \code{search_files}}
}
\value{
A logical mask volume
}
\description{
Create a preprocessing mask
}
