% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{create_preproc_mask}
\alias{create_preproc_mask}
\title{create a binary mask from pre-processed functional scans}
\usage{
create_preproc_mask(x, ...)
}
\description{
create a binary mask from pre-processed functional scans
}
