% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{dataset_name}
\alias{dataset_name}
\alias{dataset_type}
\title{Get fields from a BIDS dataset description}
\usage{
dataset_name(x, ...)

dataset_type(x, ...)
}
\arguments{
\item{x}{A \code{bids_dataset_description} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character scalar.
}
\description{
Get fields from a BIDS dataset description
}
