% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{encode}
\alias{encode}
\title{Encode a string into a BIDS key-value list}
\usage{
encode(x, ...)
}
\arguments{
\item{x}{the string to encode}

\item{...}{extra args}
}
\description{
Encode a string into a BIDS key-value list
}
