% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{flat_list}
\alias{flat_list}
\title{Get "flat" representation}
\usage{
flat_list(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
get a flattened representation of project
}
