% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmap_parser}
\alias{fmap_parser}
\title{Fieldmap parser constructor}
\usage{
fmap_parser()
}
\value{
A fieldmap BIDS parser object for parsing fieldmap files
}
\description{
Fieldmap parser constructor
}
\examples{
# Create a fieldmap parser
parser <- fmap_parser()

# Parse a fieldmap file
result <- parse(parser, "sub-01_magnitude1.nii.gz")
}
