% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{fmapspec}
\alias{fmapspec}
\title{Create a spec table for fieldmap files}
\usage{
fmapspec()
}
\description{
Create a spec table for fieldmap files
}
\value{
A list containing \code{keystruc}, \code{kinds}, and \code{type} for fieldmap files.
}
\examples{
fmapspec()
}
\keyword{internal}
