% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmriprep_anat_parser}
\alias{fmriprep_anat_parser}
\title{fMRIPrep anatomical parser constructor}
\usage{
fmriprep_anat_parser()
}
\description{
fMRIPrep anatomical parser constructor
}
\keyword{internal}
