% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{fmriprep_func_parser}
\alias{fmriprep_func_parser}
\title{construct a parser for fmriprep func types}
\usage{
fmriprep_func_parser()
}
\description{
construct a parser for fmriprep func types
}
\examples{
p <- fmriprep_func_parser()
parse(p, "sub-2002_task-mega_run-09_bold_space-MNI152NLin2009cAsym_preproc.nii.gz")
parse(p, "sub-2002_task-mega_run-08_bold_space-MNI152NLin2009cAsym_brainmask.nii.gz")
}
