% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchers.R
\name{func_parser}
\alias{func_parser}
\title{construct a parser for functional data}
\usage{
func_parser()
}
\description{
construct a parser for functional data
}
\examples{
fp <- func_parser()
parse(fp, "sub-01_ses-1_task-rest_acq-fullbrain_run-1_bold.nii.gz")
parse(fp, "sub-01_ses-1_task-nback_acq-fullbrain_run-1_events.tsv")
}
