% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids.R
\name{func_scans}
\alias{func_scans}
\alias{func_scans.bids_project}
\title{Get functional scans}
\usage{
func_scans(x, ...)

\method{func_scans}{bids_project}(
  x,
  subid = ".*",
  task = ".*",
  run = ".*",
  session = ".*",
  kind = "bold",
  full_path = TRUE,
  ...
)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}

\item{subid}{regular expression matching 'task'}

\item{task}{regular expression matching 'task'}

\item{run}{regular expression matching 'run'}

\item{session}{regular expression matching 'session'}

\item{full_path}{return full file path?}

\item{modality}{regular expression matching 'modality'}
}
\description{
extract functional scans from bids project
}
\section{Methods (by class)}{
\itemize{
\item \code{func_scans(bids_project)}: 

}}
\examples{

p <- system.file("inst/extdata/ds001", package="bidser")
fs <- func_scans(bids_project(p), subid="sub-0[123]", run="0[123]")
}
