% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids.R
\name{func_scans}
\alias{func_scans}
\alias{func_scans.bids_project}
\title{extract functional scans from bids project}
\usage{
func_scans(x, ...)

\method{func_scans}{bids_project}(
  x,
  subid = ".*",
  task = ".*",
  run = ".*",
  modality = "bold",
  full_path = TRUE,
  ...
)
}
\description{
extract functional scans from bids project
}
\section{Methods (by class)}{
\itemize{
\item \code{bids_project}: 
}}

\examples{

p <- system.file("inst/extdata/ds001", package="bidser")
fs <- func_scans(bids_project(p), subid="sub-0[123]", run="0[123]")
}
