% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{func_spec}
\alias{func_spec}
\title{func spec}
\usage{
func_spec()
}
\description{
create a spec table for "func" files
}
\details{
"spec" objects describe the pattern rules for matching bids files for a given domain, e.g. ("func", "anat", etc.)
A specification consists of two tables, a \code{keystruc} table and a \code{modalities} table. The \code{keystruc} table describes
the allowable keys (e.g. \code{sub}, \code{ses}, etc.), their order, whether they are "optional" or not, and a regular expression
indicating constraints on the pattern of the value associated with the key. The \code{name} field indicates how each key
can be referred to in the resultant parsed list structure, e.g. (\code{sub} will be named \code{subid}).

The \code{modalities} table describes the types of files and their suffixes that are allowed. the \code{modality} variable
indicates the name of the file category (e.g. "bold") and suffix is used match different file formats.
}
\keyword{internal}
