% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{funcprepspec}
\alias{funcprepspec}
\title{Create a spec table for fMRIPrep "func" files}
\usage{
funcprepspec()
}
\description{
Create a spec table for fMRIPrep "func" files
}
\value{
A list containing \code{keystruc}, \code{kinds}, and \code{type} for fMRIPrep functional files.
}
\examples{
funcprepspec()
}
\keyword{internal}
