% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{gen_parser}
\alias{gen_parser}
\title{Given a match specification, generate a parser}
\usage{
gen_parser(spec, typename = "kind")
}
\arguments{
\item{spec}{A specification object containing \code{keystruc} and \code{kinds} tables.
\code{keystruc} defines the keys and their regex patterns, and \code{kinds} defines
the possible file types and suffixes.}

\item{typename}{The name given to the final type element. Default is "kind".}
}
\value{
A parser function generated from the specification.
}
\description{
Given a match specification, generate a parser
}
\examples{
spec <- func_spec()
gen_parser(spec)
}
\keyword{internal}
