% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{gen_parser}
\alias{gen_parser}
\title{given a match specification specification, generate a parser}
\usage{
gen_parser(spec, typename = "modality")
}
\arguments{
\item{spec}{the specification as a table with \code{keystruc} and \code{modality} tables}

\item{typename}{the name given to the final type element}
}
\description{
given a match specification specification, generate a parser
}
\keyword{internal}
