% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{participants}
\alias{participants}
\title{Get participants}
\usage{
participants(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\description{
get vector of participant ids
}
\examples{
p <- system.file("inst/extdata/ds001", package="bidser")
participants(bids_project(p))
}
