% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bids_project.R
\name{plot.bids_project}
\alias{plot.bids_project}
\title{Plot a BIDS project as a dendrogram}
\usage{
\method{plot}{bids_project}(x, max_depth = Inf, ...)
}
\arguments{
\item{x}{A \code{bids_project} object}
\item{max_depth}{Maximum depth of the tree to display. Defaults to \code{Inf}.}
\item{...}{Additional arguments passed to \code{graphics::plot}.}
}
\value{
The \code{bids_project} object \code{x} is returned invisibly.
}
\description{
Visualise the hierarchical file structure of a BIDS project with a neat
dendrogram. Large projects can be truncated by setting \code{max_depth}.
}
\examples{
\donttest{
proj <- bids_project(system.file("extdata/ds001", package = "bidser"))
plot(proj)
}
}
