% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids.R
\name{preproc_scans}
\alias{preproc_scans}
\alias{preproc_scans.bids_project}
\title{extract fmriprep "preproc" scans from bids project}
\usage{
preproc_scans(x, ...)

\method{preproc_scans}{bids_project}(
  x,
  subid = ".*",
  task = ".*",
  run = ".*",
  variant = "a^",
  space = ".*",
  session = ".*",
  modality = "bold",
  full_path = FALSE,
  ...
)
}
\description{
extract fmriprep "preproc" scans from bids project
}
\section{Methods (by class)}{
\itemize{
\item \code{bids_project}: 
}}

\examples{
proj <- bids_project(system.file("inst/extdata/megalocalizer", package="bidser"), fmriprep=TRUE)
preproc_scans(proj)
}
