% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidsio.R
\name{read_confounds.bids_project}
\alias{read_confounds.bids_project}
\title{read confound files}
\usage{
\method{read_confounds}{bids_project}(
  x,
  subid = ".*",
  task = ".*",
  session = ".*",
  cvars = DEFAULT_CVARS,
  npcs = -1,
  perc_var = -1,
  nest = TRUE
)
}
\arguments{
\item{subid}{(optional) subid regex selector}

\item{task}{(optional) task regex selector}

\item{cvars}{the names of the confound variables to select. If missing, defaults to a set defined by constant \code{DEFAULT_CVARS}.}

\item{npcs}{perform pca reduction on confound matrix and reduce to \code{npcs} dimensions}

\item{perc_var}{perform pca reduction on confound matrix and retain \code{perc_var} percent of total (selected) confound variance}

\item{nest}{nest confound tables by subject/sesssion/run}
}
\description{
read in fmriprep confound tables for one or more subjects
}
