% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidsio.R
\name{read_func_scans.bids_project}
\alias{read_func_scans.bids_project}
\title{read in a set of four-dimensioanl functional scans}
\usage{
read_func_scans.bids_project(
  x,
  mask,
  mode = c("normal", "bigvec"),
  subid = "^sub-.*",
  task = ".*",
  run = ".*",
  modality = "bold",
  ...
)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{mask}{a brain mask of type \code{LogicalNeuroVol}}

\item{mode}{the file mode: 'normal' for in-memory files or 'bigvec' for on-disk files}

\item{subid}{one or more subject ids}

\item{task}{an optional task regex}

\item{run}{an optional run regex}

\item{modality}{the image modality (usually "bold")}

\item{...}{extra args send to \code{neuroim2::read_vec}}
}
\value{
a instance of type \code{NeuroVec}
}
\description{
read in a set of four-dimensioanl functional scans
}
