% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidsio.R
\name{read_preproc_scans.bids_project}
\alias{read_preproc_scans.bids_project}
\title{Read in a set of preprocessed functional scans}
\usage{
read_preproc_scans.bids_project(
  x,
  mask = NULL,
  mode = c("normal", "bigvec"),
  subid = "^sub-.*",
  task = ".*",
  run = ".*",
  modality = "bold",
  ...
)
}
\arguments{
\item{x}{A \code{bids_project} object}

\item{mask}{A brain mask of type \code{LogicalNeuroVol}, or NULL (if NULL, a mask will be created)}

\item{mode}{The file mode: 'normal' or 'bigvec'}

\item{subid}{Subject IDs (regex)}

\item{task}{Task regex}

\item{run}{Run regex}

\item{modality}{Image modality (usually "bold")}

\item{...}{Extra arguments passed to \code{neuroim2::read_vec}}
}
\value{
An instance of type \code{NeuroVec}
}
\description{
Read in a set of preprocessed functional scans
}
