% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids.R
\name{search_files}
\alias{search_files}
\alias{search_files.bids_project}
\title{search files in BIDS structure}
\usage{
search_files(x, ...)

\method{search_files}{bids_project}(x, regex = ".*", full_path = FALSE, strict = TRUE, ...)
}
\arguments{
\item{...}{additional keys to match on (e.g. subid = "01", task="wm")}

\item{regex}{a regular expression to match files}

\item{full_path}{return full_path of files}

\item{strict}{if \code{TRUE} require that a queried key must exist in match files.
Otherwise, allow matches for files missing queried key.}
}
\description{
find files in BIDS tree by matching file names
}
\examples{
 proj <- bids_project(system.file("extdata/ds001", package="bidser"), fmriprep=FALSE)
 x = search_files(proj, regex="events")
}
