% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{sessions}
\alias{sessions}
\title{Get sessions}
\usage{
sessions(x, ...)
}
\arguments{
\item{x}{the \code{bids_project} object}

\item{...}{extra args}
}
\value{
a vector of session ids
}
\description{
get a vector of sessions in the project
}
\examples{
p <- system.file("inst/extdata/ds001", package="bidser")
sessions(bids_project(p))
}
